pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rugix-bsp"
version = "0.1.0"
description = "Build Rugix Bakery BSPs from Yocto vendor BSPs."
license = "MIT OR Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[dependency-groups]
dev = ["mypy>=1.0", "ruff>=0.4", "types-pyyaml>=6.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/rugix_bsp"]

[tool.hatch.build.targets.wheel.force-include]
"layers" = "rugix_bsp/_data/layers"
"templates" = "rugix_bsp/_data/templates"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
